Split all scripts to separate modules
Added stronger type safety,
Changed tokeniser from bert to embed model
All function calls use the param_name=value syntax
