+++ .gitignore
@@ -206,4 +206,4 @@ marimo/_static/
 marimo/_lsp/
 __marimo__/
 /models/
-/v4/cache/
+/cache/
